Fix Chrome reopening
Due to a race condition between ScriptingBridge and NSWorkspace APIs, we
may have an NSRunningApplication as the active application, but a non
running SBApplication object. In Chrome, when a method is called on this
object (like windows()) ScriptingBridge will relaunch the application.
To solve this we just check if the SBApplication is running.